db: add Put method for RemoteClusterDetail
